requirements.txt
langchain==0.3.26
langchain-community>=0.0.24
langchain-google-genai
google-generativeai>=0.4.0
openai
slack_sdk
flask
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
google-cloud-aiplatform
google-cloud-storage
dateparser>=1.1.8
geopy
timezonefinder
pytz
python-dateutil
apscheduler